src/components/retrieve_from_weaviate/requirements.txt
weaviate-client==3.24.1
fondant[component]==0.7.0